requirements.txt
asgiref==3.2.3
attrs==19.3.0
autopep8==1.4.4
backcall==0.1.0
bleach==3.1.0
certifi==2019.11.28
colorama==0.4.1
decorator==4.4.1
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.1
django-cors-headers==3.2.1
django-environ==0.4.5
django-heroku==0.3.1
djangorestframework==3.11.0
entrypoints==0.3
ipykernel==5.1.3
ipython==7.11.1
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.15.2
Jinja2==2.10.3
json5==0.8.5
jsonschema==3.2.0
jupyter-client==5.3.4
jupyter-core==4.6.1
jupyterlab==1.2.4
jupyterlab-server==1.0.6
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mistune==0.8.4
mysql==0.0.2
mysqlclient==1.4.6
nbconvert==5.6.1
nbformat==5.0.3
notebook==6.0.2
pandocfilters==1.4.2
parso==0.5.2
pickleshare==0.7.5
Pillow==7.0.0
pipenv==2018.11.26
prometheus-client==0.7.1
prompt-toolkit==3.0.2
psycopg2==2.8.4
pycodestyle==2.5.0
Pygments==2.5.2
pyrsistent==0.15.7
python-dateutil==2.8.1
pytz==2019.3
pywin32==227
pywinpty==0.5.7
pyzmq==18.1.1
Send2Trash==1.5.0
six==1.13.0
sqlparse==0.3.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.3
traitlets==4.3.3
virtualenv==16.7.9
virtualenv-clone==0.5.3
virtualenvwrapper-win==1.2.5
wcwidth==0.1.8
webencodings==0.5.1
whitenoise==5.0.1

